+++ tiger-rbel/src/main/java/de/gematik/rbellogger/key/RbelKey.java
@@ -9,12 +9,13 @@ import java.security.KeyPair;
 import java.security.PrivateKey;
 import java.security.PublicKey;
 import java.util.Optional;
-import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
+import lombok.EqualsAndHashCode;
+import org.apache.commons.codec.digest.DigestUtils;
 
 @Data
-@AllArgsConstructor
+@EqualsAndHashCode(of = "hash")
 public class RbelKey {
 
   public static final int PRECEDENCE_X5C_HEADER_VALUE = 100;
@@ -23,6 +24,7 @@ public class RbelKey {
 
   private final Key key;
   private final String keyName;
+  private final byte[] hash;
 
   /**
    * The importance of this particular key. Higher value means it will be considered before
@@ -37,6 +39,7 @@ public class RbelKey {
     this.key = key;
     this.keyName = keyName;
     this.precedence = precedence;
+    this.hash = DigestUtils.sha256(key.getEncoded());
     this.matchingPublicKey = Optional.ofNullable(matchingPublicKey);
   }
 
@@ -44,7 +47,16 @@ public class RbelKey {
     this.key = key;
     this.keyName = keyName;
     this.precedence = precedence;
+    this.hash = DigestUtils.sha256(key.getEncoded());
     this.matchingPublicKey = Optional.empty();
+  }
+
+  public RbelKey(Key key, String keyName, int precedence, Optional<RbelKey> matchingPublicKey) {
+    this.key = key;
+    this.keyName = keyName;
+    this.precedence = precedence;
+    this.hash = DigestUtils.sha256(key.getEncoded());
+    this.matchingPublicKey = matchingPublicKey;
   }
 
   public Optional<KeyPair> retrieveCorrespondingKeyPair() {