+++ tiger-rbel/src/test/java/de/gematik/rbellogger/converter/VauEpa3ConverterTest.java
@@ -4,6 +4,7 @@
 
 package de.gematik.rbellogger.converter;
 
+import static de.gematik.rbellogger.testutil.RbelElementAssertion.assertThat;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import de.gematik.rbellogger.RbelLogger;
@@ -11,7 +12,6 @@ import de.gematik.rbellogger.captures.RbelFileReaderCapturer;
 import de.gematik.rbellogger.configuration.RbelConfiguration;
 import de.gematik.rbellogger.converter.initializers.RbelKeyFolderInitializer;
 import de.gematik.rbellogger.renderer.RbelHtmlRenderer;
-import de.gematik.rbellogger.testutil.RbelElementAssertion;
 import java.io.File;
 import java.nio.file.Files;
 import lombok.SneakyThrows;
@@ -49,15 +49,33 @@ public class VauEpa3ConverterTest {
   @SneakyThrows
   @Test
   void testDecryption() {
-    RbelElementAssertion.assertThat(rbelLogger.getMessageList().get(1))
+    assertThat(rbelLogger.getMessageList().get(1))
         .hasChildWithPath("$.body.AEAD_ct.decrypted_content");
-    RbelElementAssertion.assertThat(rbelLogger.getMessageList().get(2))
+    assertThat(rbelLogger.getMessageList().get(2))
         .hasChildWithPath("$.body.AEAD_ct.decrypted_content");
-    RbelElementAssertion.assertThat(rbelLogger.getMessageList().get(4))
+    assertThat(rbelLogger.getMessageList().get(4))
         .extractChildWithPath("$.body.decrypted")
         .hasStringContentEqualTo("Hello World");
-    RbelElementAssertion.assertThat(rbelLogger.getMessageList().get(5))
+    assertThat(rbelLogger.getMessageList().get(5))
         .extractChildWithPath("$.body.decrypted")
         .hasStringContentEqualTo("Right back at ya!");
+  }
+
+  @Test
+  void nestedPathProblems() throws Exception {
+    var logger =
+      RbelLogger.build(
+        new RbelConfiguration()
+          .addInitializer(new RbelKeyFolderInitializer("src/test/resources"))
+          .setActivateVauEpa3Parsing(true)
+          .addCapturer(
+            RbelFileReaderCapturer.builder()
+              .rbelFile("src/test/resources/nestedPathProblems.tgr")
+              .build()));
+    try (final var capturer = logger.getRbelCapturer()) {
+      capturer.initialize();
+    }
+    assertThat(logger.getMessageList().get(9))
+      .extractChildWithPath("$.body.decrypted.path");
   }
 }